backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.7.0
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic[email]
python-dotenv==1.0.0
resend==2.10.0
langchain-core==0.3.15
langchain-google-genai==2.0.1
langgraph==0.2.45
